Extract BreadcrumbPage component
